Separate AllocatorTest and modify stats command test
